Templates to upper case
Convert templates to upper case and expand any contractions in
english.
